pyproject.toml
[tool.poetry]
name = "pillow-test"
version = "0.1.0"
description = ""
authors = ["Bhaktij Koli <bhaktijkoli121@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
